Finishes switching of cameras with zone areas

+++ aitd/Room.cpp
@@ -24,7 +24,7 @@ void Room::load(const char *data) {
 	world_pos(1) = static_cast<float>((int16)READ_LE_UINT16(data + 6));
 	world_pos(2) = static_cast<float>((int16)READ_LE_UINT16(data + 8));
 	
-	int num_cameras = READ_LE_UINT16(data + 0xA);
+	int16 num_cameras = READ_LE_UINT16(data + 0xA);
 
 	//NOTE: different rooms share the same camera index. Some rooms are seen from two
 	// different cameras (through doors)
@@ -35,7 +35,7 @@ void Room::load(const char *data) {
 
 	//room structure data
 	const char *hard_col_data = data + READ_LE_UINT16(data);
-	int num_hard_cols = READ_LE_UINT16(hard_col_data);
+	int16 num_hard_cols = READ_LE_UINT16(hard_col_data);
 	hard_col_data += 2;
 
 	if(num_hard_cols) {	
@@ -47,7 +47,7 @@ void Room::load(const char *data) {
 
 		// Triggers
 		const char *sce_zone_data = data + READ_LE_UINT16(data + 2);
-		int num_sce_zones = READ_LE_UINT16(sce_zone_data);
+		int16 num_sce_zones = READ_LE_UINT16(sce_zone_data);
 		sce_zone_data += 2;
 	
 		for(uint32 j = 0; j < num_sce_zones; j++) {
